Fix imports and make specific imports without using the image top level module directly
